requirements.txt
fastapi==0.95.0
pandas==1.5.3
pyspellchecker==0.7.1
spacy==2.3.9
uvicorn==0.20.0